+++ Quotations Board Backend/DTOs/Bonds/YieldCurveDTO.cs
@@ -2,8 +2,8 @@ public class YieldCurveDTO
 {
     public string BondId { get; set; }
     public DateTime Date { get; set; }
-    public decimal Yield { get; set; }
+    public double Yield { get; set; }
     public decimal TotalQuotationsUsed { get; set; }
-    public decimal AverageBuyYield { get; set; }
-    public decimal AverageSellYield { get; set; }
+    public double AverageBuyYield { get; set; }
+    public double AverageSellYield { get; set; }
 }
