+++ Quotations Board Backend/DTOs/ImpliedYield/AddImpliedYieldDTO.cs
@@ -1,6 +1,6 @@
 public class AddImpliedYieldDTO
 {
     public DateTime YieldDate { get; set; }
-    public decimal Yield { get; set; }
+    public double Yield { get; set; }
     public string BondId { get; set; } = null!;
 }
