+++ Quotations Board Backend/DTOs/ImpliedYield/ImpliedYieldDTO.cs
@@ -2,7 +2,7 @@ public class ImpliedYieldDTO
 {
     public string BondName { get; set; } = null!;
     public double YearsToMaturity { get; set; }
-    public decimal Yield { get; set; }
+    public double Yield { get; set; }
     public DateTime YieldDate { get; set; }
 
 }
