+++ Quotations Board Backend/DTOs/Quotations/BondAndAverageQuotedYield.cs
@@ -3,6 +3,6 @@ public class BondAndAverageQuotedYield
     public string BondId { get; set; } = null!;
     public double BondTenor { get; set; }
     public string BondName { get; set; } = null!;
-    public decimal AverageQuotedYield { get; set; }
+    public double AverageQuotedYield { get; set; }
 
 }
