+++ Quotations Board Backend/DTOs/Quotations/QuotationAverages.cs
@@ -8,10 +8,10 @@ public class QuoteAverageData
     public string BondId { get; set; } = null!;
     public string BondIsin { get; set; } = null!;
     public string IssueNumber { get; set; } = null!;
-    public decimal AverageYield { get; set; }
-    public decimal AverageBuyVolume { get; set; }
-    public decimal AverageSellVolume { get; set; }
-    public decimal AverageSellYield { get; set; }
-    public decimal AverageBuyYield { get; set; }
-    public decimal AverageVolume { get; set; }
+    public double AverageYield { get; set; }
+    public double AverageBuyVolume { get; set; }
+    public double AverageSellVolume { get; set; }
+    public double AverageSellYield { get; set; }
+    public double AverageBuyYield { get; set; }
+    public double AverageVolume { get; set; }
 }
