+++ Quotations Board Backend/DTOs/Quotations/QuotationDTO.cs
@@ -9,12 +9,12 @@ public class Quoteinfo
     public string Id { get; set; } = null!;
     public string BondIsin { get; set; } = null!;
     public string IssueNumber { get; set; } = null!;
-    public decimal TotalBuyYield { get; set; }
-    public decimal TotalSellYield { get; set; }
-    public decimal AverageYield { get; set; }
-    public decimal TotalBuyVolume { get; set; }
-    public decimal TotalSellVolume { get; set; }
-    public decimal AverageVolume { get; set; }
+    public double TotalBuyYield { get; set; }
+    public double TotalSellYield { get; set; }
+    public double AverageYield { get; set; }
+    public double TotalBuyVolume { get; set; }
+    public double TotalSellVolume { get; set; }
+    public double AverageVolume { get; set; }
     public string InstitutionId { get; set; } = null!;
     public string BondId { get; set; } = null!;
     public DateTime CreatedAt { get; set; }
