+++ Quotations Board Backend/Models/BondTrades/BondTradeLine.cs
@@ -11,12 +11,12 @@ public class BondTradeLine
     public string BondId { get; set; } = null!;
     public string Side { get; set; } = null!;
     public string SecurityId { get; set; } = null!;
-    public decimal ExecutedSize { get; set; }
-    public decimal ExcecutedPrice { get; set; }
+    public double ExecutedSize { get; set; }
+    public double ExcecutedPrice { get; set; }
     public string ExecutionID { get; set; } = null!;
     public DateTime TransactionTime { get; set; }
     public decimal DirtyPrice { get; set; }
-    public decimal Yield { get; set; }
+    public double Yield { get; set; }
     public string? TransactionID { get; set; }
     public virtual BondTrade BondTrade { get; set; } = null!;
