+++ Quotations Board Backend/Models/Quotations/QuotationEdit.cs
@@ -6,10 +6,10 @@ public class QuotationEdit
     [Key]
     [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
     public string Id { get; set; } = null!;
-    public decimal BuyingYield { get; set; }
-    public decimal SellingYield { get; set; }
-    public decimal BuyVolume { get; set; }
-    public decimal SellVolume { get; set; }
+    public double BuyingYield { get; set; }
+    public double SellingYield { get; set; }
+    public double BuyVolume { get; set; }
+    public double SellVolume { get; set; }
     public string InstitutionId { get; set; } = null!;
     public string BondId { get; set; } = null!;
     public DateTime CreatedAt { get; set; }
